[ADD] orderService

+++ Backend/shopplyn-backend/src/main/java/com/lazxnet/e_commer/order/Dto/OrderItemResponse.java
@@ -1,5 +1,6 @@
 package com.lazxnet.e_commer.order.Dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 import com.lazxnet.e_commer.products.dto.ProductClientResponse;
@@ -10,5 +11,5 @@ public class OrderItemResponse {
     private UUID orderItemId;
     private ProductClientResponse product;
     private int quantity;
-    private Double priceAtPurchase;
+    private BigDecimal priceAtPurchase;
 }
